# [Experiment] fastdto for MakePairLabels (on top of #1734) 