Experimenting with tailored dto for https://github.com/prometheus/client_golang/pull/1734

*dto.LabelPairs are pretty bad. We got stuck with *dto.Metric in our API since the beginning. This is problematic with the notorious use of small objects and pointers everywhere and all structs are bigger to support unknown fields and parsing. We don't need any of this in client_golang - it's just overhead for us. We only need interim struct and fast encoding.

This is also the reason why we can't optimize https://github.com/prometheus/client_golang/pull/1734 further.

We always wanted to do move out, but we can't change dto.Metric use due to compatibility....

...unless we can (: 

We could start introducing it slowly and then do `Write2(m fastdto.Metric)` or so API if we want (or release v2 client golang which I would love to avoid).


Benchmarks are not ideal. I think it keeps the CPU adventage of https://github.com/prometheus/client_golang/pull/1734, but "reverts" the allocs to where they were.

```
benchstat newDesc-v1.txt newDesc-v2.txt 
goos: darwin
goarch: arm64
pkg: github.com/prometheus/client_golang/prometheus
                    │ newDesc-v1.txt │           newDesc-v2.txt            │
                    │     sec/op     │    sec/op     vs base               │
NewDesc/labels=1-2      547.5n ± 15%   420.8n ±  6%  -23.14% (p=0.002 n=6)
NewDesc/labels=3-2     1152.5n ±  3%   879.6n ± 17%  -23.68% (p=0.002 n=6)
NewDesc/labels=10-2     4.955µ ±  3%   3.929µ ±  6%  -20.71% (p=0.002 n=6)
geomean                 1.462µ         1.133µ        -22.52%

                    │ newDesc-v1.txt │           newDesc-v2.txt            │
                    │      B/op      │     B/op      vs base               │
NewDesc/labels=1-2        504.0 ± 0%     376.0 ± 0%  -25.40% (p=0.002 n=6)
NewDesc/labels=3-2       1064.0 ± 0%     680.0 ± 0%  -36.09% (p=0.002 n=6)
NewDesc/labels=10-2     4.488Ki ± 0%   3.301Ki ± 0%  -26.46% (p=0.002 n=6)
geomean                 1.319Ki          952.5       -29.48%

                    │ newDesc-v1.txt │          newDesc-v2.txt           │
                    │   allocs/op    │ allocs/op   vs base               │
NewDesc/labels=1-2        15.00 ± 0%   10.00 ± 0%  -33.33% (p=0.002 n=6)
NewDesc/labels=3-2        27.00 ± 0%   12.00 ± 0%  -55.56% (p=0.002 n=6)
NewDesc/labels=10-2       72.00 ± 0%   22.00 ± 0%  -69.44% (p=0.002 n=6)
geomean                   30.78        13.82       -55.10%
```

```
benchstat makeLabelPairs-v1.txt makeLabelPairs-v2.txt
goos: darwin
goarch: arm64
pkg: github.com/prometheus/client_golang/prometheus
                           │ makeLabelPairs-v1.txt │       makeLabelPairs-v2.txt        │
                           │        sec/op         │    sec/op     vs base              │
MakeLabelPairs/labels=1-2             65.21n ±  6%   66.42n ± 19%       ~ (p=0.132 n=6)
MakeLabelPairs/labels=3-2             156.8n ±  1%   171.2n ±  3%  +9.18% (p=0.002 n=6)
MakeLabelPairs/labels=10-2            501.5n ± 12%   516.2n ±  4%       ~ (p=0.310 n=6)
geomean                               172.4n         180.4n        +4.61%

                           │ makeLabelPairs-v1.txt │       makeLabelPairs-v2.txt        │
                           │         B/op          │    B/op      vs base               │
MakeLabelPairs/labels=1-2               96.00 ± 0%   144.00 ± 0%  +50.00% (p=0.002 n=6)
MakeLabelPairs/labels=3-2               288.0 ± 0%    432.0 ± 0%  +50.00% (p=0.002 n=6)
MakeLabelPairs/labels=10-2              960.0 ± 0%   1440.0 ± 0%  +50.00% (p=0.002 n=6)
geomean                                 298.3         447.4       +50.00%

                           │ makeLabelPairs-v1.txt │       makeLabelPairs-v2.txt        │
                           │       allocs/op       │ allocs/op   vs base                │
MakeLabelPairs/labels=1-2               3.000 ± 0%   3.000 ± 0%       ~ (p=1.000 n=6) ¹
MakeLabelPairs/labels=3-2               7.000 ± 0%   7.000 ± 0%       ~ (p=1.000 n=6) ¹
MakeLabelPairs/labels=10-2              21.00 ± 0%   21.00 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                 7.612        7.612       +0.00%
¹ all samples are equal
```

I think for the next iteration we could try moving this cost to `Write` (so when collect is used). When collecting one could pool/reuse dto.Metrics OR use `Write2` with fastdto Metric 🤔